Finish dialog working

+++ app/src/main/java/com/kent/lw/brainendurancetrainingmobileapp/MapHelper.java
@@ -28,8 +28,8 @@ public class MapHelper {
         return new LatLng(location.getLatitude(), location.getLongitude());
     }
 
-    public double getDistance(LatLng l1, LatLng l2) {
-        return SphericalUtil.computeDistanceBetween(l1, l2);
+    public float getDistance(LatLng l1, LatLng l2) {
+        return (float) SphericalUtil.computeDistanceBetween(l1, l2);
     }
 
     public void removePolylines(List<Polyline> polylineList) {
